feat(notify,screen): drop persistent monitor notification; invert 3-finger scroll
- Replace the foreground AgentMonitorService (which required a persistent
  "watching the agent" notification) with a process-scope AgentMonitor started
  from AntigravityApp. Now the app only posts a notification when the agent needs
  input, finishes, errors, or a quota runs low — no ongoing notification. Removes
  the foreground-service permissions and service declaration.
- Invert the 3-finger scroll direction (swipe down scrolls content up).
